Refactor SleepResult to use SystemSleepResult everywhere.
