+++ hal/inc/hal_platform_compat.h
@@ -56,7 +56,7 @@
 #if PLATFORM_ID == 6 || PLATFORM_ID == 8 || PLATFORM_ID == 10
 #define HAL_PLATFORM_DCT 1
 #define HAL_PLATFORM_RNG 1
-#define HAL_PLATFORM_SLEEP_2_0 1
+#define HAL_PLATFORM_SLEEP20 1
 #endif
 
 #define HAL_PLATFORM_NCP 	(0)
