+++ hal/src/argon/hal_platform_config.h
@@ -7,7 +7,7 @@
 #define HAL_PLATFORM_NCP_AT (1)
 #define HAL_PLATFORM_NCP_UPDATABLE (1)
 #define HAL_PLATFORM_WIFI (1)
-#define HAL_PLATFORM_SLEEP_2_0 (1)
+#define HAL_PLATFORM_SLEEP20 (1)
 
 #define HAL_PLATFORM_RADIO_ANTENNA_INTERNAL (1)
 #define HAL_PLATFORM_RADIO_ANTENNA_EXTERNAL (1)
