+++ hal/src/boron/hal_platform_config.h
@@ -7,7 +7,7 @@
 #define HAL_PLATFORM_NCP_AT (1)
 #define HAL_PLATFORM_CELLULAR (1)
 #define HAL_PLATFORM_SETUP_BUTTON_UX (1)
-#define HAL_PLATFORM_SLEEP_2_0 (1)
+#define HAL_PLATFORM_SLEEP20 (1)
 
 #if PLATFORM_ID == PLATFORM_BORON
