+++ hal/src/stm32f2xx/sleep_hal.cpp
@@ -17,7 +17,7 @@
 
 #include "sleep_hal.h"
 
-#if HAL_PLATFORM_SLEEP_2_0
+#if HAL_PLATFORM_SLEEP20
 
 #include "stm32f2xx.h"
 #include "gpio_hal.h"
@@ -361,4 +361,4 @@ int hal_sleep_enter(const hal_sleep_config_t* config, hal_wakeup_source_base_t**
     return ret;
 }
 
-#endif // HAL_PLATFORM_SLEEP_2_0
+#endif // HAL_PLATFORM_SLEEP20
