+++ hal/src/xenon/hal_platform_config.h
@@ -4,7 +4,7 @@
 #include "platforms.h"
 
 #define HAL_PLATFORM_USART2 (1)
-#define HAL_PLATFORM_SLEEP_2_0 (1)
+#define HAL_PLATFORM_SLEEP20 (1)
 
 #define HAL_PLATFORM_RADIO_ANTENNA_INTERNAL (1)
 #define HAL_PLATFORM_RADIO_ANTENNA_EXTERNAL (1)
